requirements.txt
langchain
python-dotenv
streamlit
huggingface_hub
